requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
altgraph==0.17.4
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
macholib==1.16.3
modulegraph==0.19.6
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
py2app==0.28.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pyparsing==3.2.3
python-telegram-bot==22.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
twilio==9.6.4
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
